Support implicitly-references-foreign-primary-key syntax.
In <https://github.com/dgoffredo/sqliteviz/issues/1>, github user dkg
pointed out that sqlite's "references some_table" syntax is not
supported by this tool. SQLite understands that syntax to mean
"references some_table(the, primary, key, of, some_table)", but the
pragma_foreign_key_list(...) table function does not encode that
understanding. Instead, the "to column" field is null.

This revision deduces the corresponding "to column" in those cases
by keeping track of the primary keys of all tables, and by using the
"seq" column of the pragma_foreign_key_list(...) table, which allows us
to find the foreign primary key column that corresponds to the
referencing foreign key column.

In playing with this, I also noticed that we weren't supporting default
column types, i.e. no type or dynamic type. Now untyped columns receive
the "any" type name (which is not a real thing in SQL, but looks good in
the graphviz output).

dkg's example of a schema that shows the original issue, and another
that I was playing with, are now included as examples in addition to the
`schema.sql` mentioned in the readme.